Make evaluation of SetRef lazy again
The modifyRef test case doesn't need SetRef to be strict anymore now that
unsafeFreezeRef has been made strict.

But in fact, SetRef *should* be strict in order to match the semantics of the
generated C code. The reason why I'm making it lazy is that this is more consistent
because there are many other cases in which evaluation is too lazy.